+++ src/Aws/Route53/Resources/route53-2013-04-01.php
@@ -66,6 +66,73 @@ return array (
         ),
     ),
     'operations' => array(
+        'AssociateVPCWithHostedZone' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/2013-04-01/hostedzone/{HostedZoneId}/associatevpc',
+            'class' => 'Guzzle\\Service\\Command\\OperationCommand',
+            'responseClass' => 'AssociateVPCWithHostedZoneResponse',
+            'responseType' => 'model',
+            'data' => array(
+                'xmlRoot' => array(
+                    'name' => 'AssociateVPCWithHostedZoneRequest',
+                    'namespaces' => array(
+                        'https://route53.amazonaws.com/doc/2013-04-01/',
+                    ),
+                ),
+            ),
+            'parameters' => array(
+                'HostedZoneId' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'uri',
+                    'maxLength' => 32,
+                ),
+                'VPC' => array(
+                    'required' => true,
+                    'type' => 'object',
+                    'location' => 'xml',
+                    'properties' => array(
+                        'VPCRegion' => array(
+                            'type' => 'string',
+                            'minLength' => 1,
+                            'maxLength' => 64,
+                        ),
+                        'VPCId' => array(
+                            'type' => 'string',
+                            'maxLength' => 1024,
+                        ),
+                    ),
+                ),
+                'Comment' => array(
+                    'type' => 'string',
+                    'location' => 'xml',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/xml',
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'class' => 'NoSuchHostedZoneException',
+                ),
+                array(
+                    'reason' => 'The hosted zone you are trying to create for your VPC_ID does not belong to you. Route 53 returns this error when the VPC specified by VPCId does not belong to you.',
+                    'class' => 'InvalidVPCIdException',
+                ),
+                array(
+                    'reason' => 'Some value specified in the request is invalid or the XML document is malformed.',
+                    'class' => 'InvalidInputException',
+                ),
+                array(
+                    'reason' => 'The hosted zone you are trying to associate VPC with doesn\'t have any VPC association. Route 53 currently doesn\'t support associate a VPC with a public hosted zone.',
+                    'class' => 'PublicZoneVPCAssociationException',
+                ),
+                array(
+                    'class' => 'ConflictingDomainExistsException',
+                ),
+            ),
+        ),
         'ChangeResourceRecordSets' => array(
             'httpMethod' => 'POST',
             'uri' => '/2013-04-01/hostedzone/{HostedZoneId}/rrset/',
@@ -418,6 +485,21 @@ return array (
                     'location' => 'xml',
                     'maxLength' => 1024,
                 ),
+                'VPC' => array(
+                    'type' => 'object',
+                    'location' => 'xml',
+                    'properties' => array(
+                        'VPCRegion' => array(
+                            'type' => 'string',
+                            'minLength' => 1,
+                            'maxLength' => 64,
+                        ),
+                        'VPCId' => array(
+                            'type' => 'string',
+                            'maxLength' => 1024,
+                        ),
+                    ),
+                ),
                 'CallerReference' => array(
                     'required' => true,
                     'type' => 'string',
@@ -433,7 +515,16 @@ return array (
                             'type' => 'string',
                             'maxLength' => 256,
                         ),
+                        'PrivateZone' => array(
+                            'type' => 'boolean',
+                            'format' => 'boolean-string',
+                        ),
                     ),
+                ),
+                'DelegationSetId' => array(
+                    'type' => 'string',
+                    'location' => 'xml',
+                    'maxLength' => 32,
                 ),
                 'command.expects' => array(
                     'static' => true,
@@ -454,12 +545,90 @@ return array (
                     'class' => 'TooManyHostedZonesException',
                 ),
                 array(
+                    'reason' => 'The hosted zone you are trying to create for your VPC_ID does not belong to you. Route 53 returns this error when the VPC specified by VPCId does not belong to you.',
+                    'class' => 'InvalidVPCIdException',
+                ),
+                array(
                     'reason' => 'Some value specified in the request is invalid or the XML document is malformed.',
                     'class' => 'InvalidInputException',
                 ),
                 array(
                     'reason' => 'Route 53 allows some duplicate domain names, but there is a maximum number of duplicate names. This error indicates that you have reached that maximum. If you want to create another hosted zone with the same name and Route 53 generates this error, you can request an increase to the limit on the Contact Us page.',
                     'class' => 'DelegationSetNotAvailableException',
+                ),
+                array(
+                    'class' => 'ConflictingDomainExistsException',
+                ),
+                array(
+                    'reason' => 'The specified delegation set does not exist.',
+                    'class' => 'NoSuchDelegationSetException',
+                ),
+                array(
+                    'reason' => 'The specified delegation set has not been marked as reusable.',
+                    'class' => 'DelegationSetNotReusableException',
+                ),
+            ),
+        ),
+        'CreateReusableDelegationSet' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/2013-04-01/delegationset',
+            'class' => 'Guzzle\\Service\\Command\\OperationCommand',
+            'responseClass' => 'CreateReusableDelegationSetResponse',
+            'responseType' => 'model',
+            'data' => array(
+                'xmlRoot' => array(
+                    'name' => 'CreateReusableDelegationSetRequest',
+                    'namespaces' => array(
+                        'https://route53.amazonaws.com/doc/2013-04-01/',
+                    ),
+                ),
+            ),
+            'parameters' => array(
+                'CallerReference' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'xml',
+                    'minLength' => 1,
+                    'maxLength' => 128,
+                ),
+                'HostedZoneId' => array(
+                    'type' => 'string',
+                    'location' => 'xml',
+                    'maxLength' => 32,
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/xml',
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'A delegation set with the same owner and caller reference combination has already been created.',
+                    'class' => 'DelegationSetAlreadyCreatedException',
+                ),
+                array(
+                    'reason' => 'The limits specified for a resource have been exceeded.',
+                    'class' => 'LimitsExceededException',
+                ),
+                array(
+                    'reason' => 'The specified HostedZone cannot be found.',
+                    'class' => 'HostedZoneNotFoundException',
+                ),
+                array(
+                    'reason' => 'At least one of the specified arguments is invalid.',
+                    'class' => 'InvalidArgumentException',
+                ),
+                array(
+                    'reason' => 'Some value specified in the request is invalid or the XML document is malformed.',
+                    'class' => 'InvalidInputException',
+                ),
+                array(
+                    'reason' => 'Route 53 allows some duplicate domain names, but there is a maximum number of duplicate names. This error indicates that you have reached that maximum. If you want to create another hosted zone with the same name and Route 53 generates this error, you can request an increase to the limit on the Contact Us page.',
+                    'class' => 'DelegationSetNotAvailableException',
+                ),
+                array(
+                    'reason' => 'The specified delegation set has already been marked as reusable.',
+                    'class' => 'DelegationSetAlreadyReusableException',
                 ),
             ),
         ),
@@ -524,6 +693,107 @@ return array (
                 array(
                     'reason' => 'The request was rejected because Route 53 was still processing a prior request.',
                     'class' => 'PriorRequestNotCompleteException',
+                ),
+                array(
+                    'reason' => 'Some value specified in the request is invalid or the XML document is malformed.',
+                    'class' => 'InvalidInputException',
+                ),
+            ),
+        ),
+        'DeleteReusableDelegationSet' => array(
+            'httpMethod' => 'DELETE',
+            'uri' => '/2013-04-01/delegationset/{Id}',
+            'class' => 'Guzzle\\Service\\Command\\OperationCommand',
+            'responseClass' => 'DeleteReusableDelegationSetResponse',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Id' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'uri',
+                    'maxLength' => 32,
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'The specified delegation set does not exist.',
+                    'class' => 'NoSuchDelegationSetException',
+                ),
+                array(
+                    'reason' => 'The specified delegation contains associated hosted zones which must be deleted before the reusable delegation set can be deleted.',
+                    'class' => 'DelegationSetInUseException',
+                ),
+                array(
+                    'reason' => 'The specified delegation set has not been marked as reusable.',
+                    'class' => 'DelegationSetNotReusableException',
+                ),
+                array(
+                    'reason' => 'Some value specified in the request is invalid or the XML document is malformed.',
+                    'class' => 'InvalidInputException',
+                ),
+            ),
+        ),
+        'DisassociateVPCFromHostedZone' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/2013-04-01/hostedzone/{HostedZoneId}/disassociatevpc',
+            'class' => 'Guzzle\\Service\\Command\\OperationCommand',
+            'responseClass' => 'DisassociateVPCFromHostedZoneResponse',
+            'responseType' => 'model',
+            'data' => array(
+                'xmlRoot' => array(
+                    'name' => 'DisassociateVPCFromHostedZoneRequest',
+                    'namespaces' => array(
+                        'https://route53.amazonaws.com/doc/2013-04-01/',
+                    ),
+                ),
+            ),
+            'parameters' => array(
+                'HostedZoneId' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'uri',
+                    'maxLength' => 32,
+                ),
+                'VPC' => array(
+                    'required' => true,
+                    'type' => 'object',
+                    'location' => 'xml',
+                    'properties' => array(
+                        'VPCRegion' => array(
+                            'type' => 'string',
+                            'minLength' => 1,
+                            'maxLength' => 64,
+                        ),
+                        'VPCId' => array(
+                            'type' => 'string',
+                            'maxLength' => 1024,
+                        ),
+                    ),
+                ),
+                'Comment' => array(
+                    'type' => 'string',
+                    'location' => 'xml',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/xml',
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'class' => 'NoSuchHostedZoneException',
+                ),
+                array(
+                    'reason' => 'The hosted zone you are trying to create for your VPC_ID does not belong to you. Route 53 returns this error when the VPC specified by VPCId does not belong to you.',
+                    'class' => 'InvalidVPCIdException',
+                ),
+                array(
+                    'reason' => 'The VPC you specified is not currently associated with the hosted zone.',
+                    'class' => 'VPCAssociationNotFoundException',
+                ),
+                array(
+                    'reason' => 'The VPC you are trying to disassociate from the hosted zone is the last the VPC that is associated with the hosted zone. Route 53 currently doesn\'t support disassociate the last VPC from the hosted zone.',
+                    'class' => 'LastVPCAssociationException',
                 ),
                 array(
                     'reason' => 'Some value specified in the request is invalid or the XML document is malformed.',
@@ -665,6 +935,56 @@ return array (
                 ),
             ),
         ),
+        'GetHealthCheckLastFailureReason' => array(
+            'httpMethod' => 'GET',
+            'uri' => '/2013-04-01/healthcheck/{HealthCheckId}/lastfailurereason',
+            'class' => 'Guzzle\\Service\\Command\\OperationCommand',
+            'responseClass' => 'GetHealthCheckLastFailureReasonResponse',
+            'responseType' => 'model',
+            'parameters' => array(
+                'HealthCheckId' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'uri',
+                    'maxLength' => 64,
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/xml',
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'The health check you are trying to get or delete does not exist.',
+                    'class' => 'NoSuchHealthCheckException',
+                ),
+            ),
+        ),
+        'GetHealthCheckStatus' => array(
+            'httpMethod' => 'GET',
+            'uri' => '/2013-04-01/healthcheck/{HealthCheckId}/status',
+            'class' => 'Guzzle\\Service\\Command\\OperationCommand',
+            'responseClass' => 'GetHealthCheckStatusResponse',
+            'responseType' => 'model',
+            'parameters' => array(
+                'HealthCheckId' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'uri',
+                    'maxLength' => 64,
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/xml',
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'The health check you are trying to get or delete does not exist.',
+                    'class' => 'NoSuchHealthCheckException',
+                ),
+            ),
+        ),
         'GetHostedZone' => array(
             'httpMethod' => 'GET',
             'uri' => '/2013-04-01/hostedzone/{Id}',
@@ -689,6 +1009,39 @@ return array (
             'errorResponses' => array(
                 array(
                     'class' => 'NoSuchHostedZoneException',
+                ),
+                array(
+                    'reason' => 'Some value specified in the request is invalid or the XML document is malformed.',
+                    'class' => 'InvalidInputException',
+                ),
+            ),
+        ),
+        'GetReusableDelegationSet' => array(
+            'httpMethod' => 'GET',
+            'uri' => '/2013-04-01/delegationset/{Id}',
+            'class' => 'Guzzle\\Service\\Command\\OperationCommand',
+            'responseClass' => 'GetReusableDelegationSetResponse',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Id' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'uri',
+                    'maxLength' => 32,
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/xml',
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'The specified delegation set does not exist.',
+                    'class' => 'NoSuchDelegationSetException',
+                ),
+                array(
+                    'reason' => 'The specified delegation set has not been marked as reusable.',
+                    'class' => 'DelegationSetNotReusableException',
                 ),
                 array(
                     'reason' => 'Some value specified in the request is invalid or the XML document is malformed.',
@@ -793,6 +1146,12 @@ return array (
                     'location' => 'query',
                     'sentAs' => 'maxitems',
                 ),
+                'DelegationSetId' => array(
+                    'type' => 'string',
+                    'location' => 'query',
+                    'sentAs' => 'delegationsetid',
+                    'maxLength' => 32,
+                ),
                 'command.expects' => array(
                     'static' => true,
                     'default' => 'application/xml',
@@ -802,6 +1161,14 @@ return array (
                 array(
                     'reason' => 'Some value specified in the request is invalid or the XML document is malformed.',
                     'class' => 'InvalidInputException',
+                ),
+                array(
+                    'reason' => 'The specified delegation set does not exist.',
+                    'class' => 'NoSuchDelegationSetException',
+                ),
+                array(
+                    'reason' => 'The specified delegation set has not been marked as reusable.',
+                    'class' => 'DelegationSetNotReusableException',
                 ),
             ),
         ),
@@ -853,6 +1220,36 @@ return array (
                 array(
                     'class' => 'NoSuchHostedZoneException',
                 ),
+                array(
+                    'reason' => 'Some value specified in the request is invalid or the XML document is malformed.',
+                    'class' => 'InvalidInputException',
+                ),
+            ),
+        ),
+        'ListReusableDelegationSets' => array(
+            'httpMethod' => 'GET',
+            'uri' => '/2013-04-01/delegationset',
+            'class' => 'Guzzle\\Service\\Command\\OperationCommand',
+            'responseClass' => 'ListReusableDelegationSetsResponse',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Marker' => array(
+                    'type' => 'string',
+                    'location' => 'query',
+                    'sentAs' => 'marker',
+                    'maxLength' => 64,
+                ),
+                'MaxItems' => array(
+                    'type' => 'string',
+                    'location' => 'query',
+                    'sentAs' => 'maxitems',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/xml',
+                ),
+            ),
+            'errorResponses' => array(
                 array(
                     'reason' => 'Some value specified in the request is invalid or the XML document is malformed.',
                     'class' => 'InvalidInputException',
@@ -1034,6 +1431,34 @@ return array (
         ),
     ),
     'models' => array(
+        'AssociateVPCWithHostedZoneResponse' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'ChangeInfo' => array(
+                    'type' => 'object',
+                    'location' => 'xml',
+                    'properties' => array(
+                        'Id' => array(
+                            'type' => 'string',
+                        ),
+                        'Status' => array(
+                            'type' => 'string',
+                        ),
+                        'SubmittedAt' => array(
+                            'type' => 'string',
+                        ),
+                        'Comment' => array(
+                            'type' => 'string',
+                        ),
+                    ),
+                ),
+                'RequestId' => array(
+                    'location' => 'header',
+                    'sentAs' => 'x-amz-request-id',
+                ),
+            ),
+        ),
         'ChangeResourceRecordSetsResponse' => array(
             'type' => 'object',
             'additionalProperties' => true,
@@ -1153,6 +1578,9 @@ return array (
                                 'Comment' => array(
                                     'type' => 'string',
                                 ),
+                                'PrivateZone' => array(
+                                    'type' => 'boolean',
+                                ),
                             ),
                         ),
                         'ResourceRecordSetCount' => array(
@@ -1182,6 +1610,58 @@ return array (
                     'type' => 'object',
                     'location' => 'xml',
                     'properties' => array(
+                        'Id' => array(
+                            'type' => 'string',
+                        ),
+                        'CallerReference' => array(
+                            'type' => 'string',
+                        ),
+                        'NameServers' => array(
+                            'type' => 'array',
+                            'items' => array(
+                                'name' => 'NameServer',
+                                'type' => 'string',
+                                'sentAs' => 'NameServer',
+                            ),
+                        ),
+                    ),
+                ),
+                'VPC' => array(
+                    'type' => 'object',
+                    'location' => 'xml',
+                    'properties' => array(
+                        'VPCRegion' => array(
+                            'type' => 'string',
+                        ),
+                        'VPCId' => array(
+                            'type' => 'string',
+                        ),
+                    ),
+                ),
+                'Location' => array(
+                    'type' => 'string',
+                    'location' => 'header',
+                ),
+                'RequestId' => array(
+                    'location' => 'header',
+                    'sentAs' => 'x-amz-request-id',
+                ),
+            ),
+        ),
+        'CreateReusableDelegationSetResponse' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'DelegationSet' => array(
+                    'type' => 'object',
+                    'location' => 'xml',
+                    'properties' => array(
+                        'Id' => array(
+                            'type' => 'string',
+                        ),
+                        'CallerReference' => array(
+                            'type' => 'string',
+                        ),
                         'NameServers' => array(
                             'type' => 'array',
                             'items' => array(
@@ -1213,6 +1693,44 @@ return array (
             ),
         ),
         'DeleteHostedZoneResponse' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'ChangeInfo' => array(
+                    'type' => 'object',
+                    'location' => 'xml',
+                    'properties' => array(
+                        'Id' => array(
+                            'type' => 'string',
+                        ),
+                        'Status' => array(
+                            'type' => 'string',
+                        ),
+                        'SubmittedAt' => array(
+                            'type' => 'string',
+                        ),
+                        'Comment' => array(
+                            'type' => 'string',
+                        ),
+                    ),
+                ),
+                'RequestId' => array(
+                    'location' => 'header',
+                    'sentAs' => 'x-amz-request-id',
+                ),
+            ),
+        ),
+        'DeleteReusableDelegationSetResponse' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'RequestId' => array(
+                    'location' => 'header',
+                    'sentAs' => 'x-amz-request-id',
+                ),
+            ),
+        ),
+        'DisassociateVPCFromHostedZoneResponse' => array(
             'type' => 'object',
             'additionalProperties' => true,
             'properties' => array(
@@ -1388,6 +1906,76 @@ return array (
                 ),
             ),
         ),
+        'GetHealthCheckLastFailureReasonResponse' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'HealthCheckObservations' => array(
+                    'type' => 'array',
+                    'location' => 'xml',
+                    'items' => array(
+                        'name' => 'HealthCheckObservation',
+                        'type' => 'object',
+                        'sentAs' => 'HealthCheckObservation',
+                        'properties' => array(
+                            'IPAddress' => array(
+                                'type' => 'string',
+                            ),
+                            'StatusReport' => array(
+                                'type' => 'object',
+                                'properties' => array(
+                                    'Status' => array(
+                                        'type' => 'string',
+                                    ),
+                                    'CheckedTime' => array(
+                                        'type' => 'string',
+                                    ),
+                                ),
+                            ),
+                        ),
+                    ),
+                ),
+                'RequestId' => array(
+                    'location' => 'header',
+                    'sentAs' => 'x-amz-request-id',
+                ),
+            ),
+        ),
+        'GetHealthCheckStatusResponse' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'HealthCheckObservations' => array(
+                    'type' => 'array',
+                    'location' => 'xml',
+                    'items' => array(
+                        'name' => 'HealthCheckObservation',
+                        'type' => 'object',
+                        'sentAs' => 'HealthCheckObservation',
+                        'properties' => array(
+                            'IPAddress' => array(
+                                'type' => 'string',
+                            ),
+                            'StatusReport' => array(
+                                'type' => 'object',
+                                'properties' => array(
+                                    'Status' => array(
+                                        'type' => 'string',
+                                    ),
+                                    'CheckedTime' => array(
+                                        'type' => 'string',
+                                    ),
+                                ),
+                            ),
+                        ),
+                    ),
+                ),
+                'RequestId' => array(
+                    'location' => 'header',
+                    'sentAs' => 'x-amz-request-id',
+                ),
+            ),
+        ),
         'GetHostedZoneResponse' => array(
             'type' => 'object',
             'additionalProperties' => true,
@@ -1411,6 +1999,9 @@ return array (
                                 'Comment' => array(
                                     'type' => 'string',
                                 ),
+                                'PrivateZone' => array(
+                                    'type' => 'boolean',
+                                ),
                             ),
                         ),
                         'ResourceRecordSetCount' => array(
@@ -1422,6 +2013,59 @@ return array (
                     'type' => 'object',
                     'location' => 'xml',
                     'properties' => array(
+                        'Id' => array(
+                            'type' => 'string',
+                        ),
+                        'CallerReference' => array(
+                            'type' => 'string',
+                        ),
+                        'NameServers' => array(
+                            'type' => 'array',
+                            'items' => array(
+                                'name' => 'NameServer',
+                                'type' => 'string',
+                                'sentAs' => 'NameServer',
+                            ),
+                        ),
+                    ),
+                ),
+                'VPCs' => array(
+                    'type' => 'array',
+                    'location' => 'xml',
+                    'items' => array(
+                        'name' => 'VPC',
+                        'type' => 'object',
+                        'sentAs' => 'VPC',
+                        'properties' => array(
+                            'VPCRegion' => array(
+                                'type' => 'string',
+                            ),
+                            'VPCId' => array(
+                                'type' => 'string',
+                            ),
+                        ),
+                    ),
+                ),
+                'RequestId' => array(
+                    'location' => 'header',
+                    'sentAs' => 'x-amz-request-id',
+                ),
+            ),
+        ),
+        'GetReusableDelegationSetResponse' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'DelegationSet' => array(
+                    'type' => 'object',
+                    'location' => 'xml',
+                    'properties' => array(
+                        'Id' => array(
+                            'type' => 'string',
+                        ),
+                        'CallerReference' => array(
+                            'type' => 'string',
+                        ),
                         'NameServers' => array(
                             'type' => 'array',
                             'items' => array(
@@ -1599,6 +2243,9 @@ return array (
                                     'Comment' => array(
                                         'type' => 'string',
                                     ),
+                                    'PrivateZone' => array(
+                                        'type' => 'boolean',
+                                    ),
                                 ),
                             ),
                             'ResourceRecordSetCount' => array(
@@ -1722,6 +2369,57 @@ return array (
                     'location' => 'xml',
                 ),
                 'NextRecordIdentifier' => array(
+                    'type' => 'string',
+                    'location' => 'xml',
+                ),
+                'MaxItems' => array(
+                    'type' => 'string',
+                    'location' => 'xml',
+                ),
+                'RequestId' => array(
+                    'location' => 'header',
+                    'sentAs' => 'x-amz-request-id',
+                ),
+            ),
+        ),
+        'ListReusableDelegationSetsResponse' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'DelegationSets' => array(
+                    'type' => 'array',
+                    'location' => 'xml',
+                    'items' => array(
+                        'name' => 'DelegationSet',
+                        'type' => 'object',
+                        'sentAs' => 'DelegationSet',
+                        'properties' => array(
+                            'Id' => array(
+                                'type' => 'string',
+                            ),
+                            'CallerReference' => array(
+                                'type' => 'string',
+                            ),
+                            'NameServers' => array(
+                                'type' => 'array',
+                                'items' => array(
+                                    'name' => 'NameServer',
+                                    'type' => 'string',
+                                    'sentAs' => 'NameServer',
+                                ),
+                            ),
+                        ),
+                    ),
+                ),
+                'Marker' => array(
+                    'type' => 'string',
+                    'location' => 'xml',
+                ),
+                'IsTruncated' => array(
+                    'type' => 'boolean',
+                    'location' => 'xml',
+                ),
+                'NextMarker' => array(
                     'type' => 'string',
                     'location' => 'xml',
                 ),
@@ -1900,6 +2598,13 @@ return array (
                 'NextRecordType',
                 'NextRecordIdentifier',
             ),
+        ),
+        'ListReusableDelegationSets' => array(
+            'input_token' => 'Marker',
+            'output_token' => 'NextMarker',
+            'more_results' => 'IsTruncated',
+            'limit_key' => 'MaxItems',
+            'result_key' => 'DelegationSets',
         ),
     ),
 );